feat: PC에서 채널 목록 접속했을때 최근 메세지 상세페이지로 리다이렉트 하기
