+++ src/app/(providers)/(root)/[workspaceId]/channels/_hooks/useChannelQuery.ts
@@ -2,13 +2,10 @@ import { useQuery } from '@tanstack/react-query';
 import { QUERY_KEYS } from '../_constants/constants';
 import api from '@/api';
 import { GetSearchWorkspaceUsersProps } from '@/types/workspaceUser';
+import { getChannelsOptions } from '../_utils/getQueryOptions';
 
 export const useGetChannels = (workspaceId: number) => {
-  return useQuery({
-    queryKey: QUERY_KEYS.CHANNELS(workspaceId),
-    queryFn: api.channel.getChannels,
-    refetchOnWindowFocus: false
-  });
+  return useQuery(getChannelsOptions(workspaceId));
 };
 
 export const useGetSearchWorkspaceUsers = ({ workspace_id, term }: GetSearchWorkspaceUsersProps) => {
