+++ src/app/(providers)/(root)/[workspaceId]/channels/_utils/getQueryOptions.ts
@@ -1,0 +1,11 @@
+import { queryOptions } from '@tanstack/react-query';
+import { QUERY_KEYS } from '../_constants/constants';
+import api from '@/api';
+
+export const getChannelsOptions = (workspaceId: number) => {
+  return queryOptions({
+    queryKey: QUERY_KEYS.CHANNELS(workspaceId),
+    queryFn: api.channel.getChannels,
+    refetchOnWindowFocus: false
+  });
+};
