+++ src/app/(providers)/(root)/[workspaceId]/channels/layout.tsx
@@ -2,7 +2,7 @@
 
 import { StrictPropsWithChildren } from '@/types/common';
 import { ResponseContainer, ResponseContent } from './_components/ResponseLayout';
-import useIsPC from '@/hooks/useIsPc';
+import useIsPC from '@/hooks/useIsPC';
 
 const ChannelListLayout = ({ list, children }: StrictPropsWithChildren<{ list: React.ReactNode }>) => {
   const isPC = useIsPC();
